requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-auditlog==3.3.0
ipython==9.6.0
django-rosetta==0.10.3
django-cors-headers==4.9.0
environs==14.5.0
gunicorn==23.0.0
whitenoise==6.11.0
dj-database-url==3.0.1
psycopg2==2.9.11
celery==5.6.1
requests==2.32.5
lxml==6.0.2
django-redis==6.0.0
django-filter==25.2
